reviewed

+++ Oracle.19.Managed/Tests/xUpdate/TruncateTableTests/Tests.xUpdate.TruncateTableTests.TruncateTableTest(Oracle.19.Managed).sql
@@ -15,8 +15,8 @@ BeforeExecute
 
 CREATE TABLE "TestTrun"
 (
-	ID       Int     NOT NULL,
-	"Field1" Decimal NOT NULL,
+	ID       Int             NOT NULL,
+	"Field1" Decimal(28, 10) NOT NULL,
 
 	CONSTRAINT "PK_TestTrun" PRIMARY KEY (ID)
 )
